requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post7
sortedcontainers==2.4.0
threadpoolctl==3.2.0
pyDOE==0.3.8
pymoo==0.4.2
tensorflow==2.12.0
